Updated args
This set of changes updates the arg list to reflect all the various arguments that the histogram (and the inverted index) can take.  It allows for the implementation of stop file and reading a single file from the command line.
Using the other args will result in an error message.
This is the first step to implementing the various arguments one by one.